refactor: Merge `ItemSet`, `NonTerminal`, and `Production` into `parser_types`. (#97)
