+++ src/log_surgeon/parser_types.cpp
@@ -1,8 +1,5 @@
-#include <cstdint>
-
 #include <log_surgeon/Constants.hpp>
-#include <log_surgeon/NonTerminal.hpp>
-#include <log_surgeon/Production.hpp>
+#include <log_surgeon/parser_types.hpp>
 
 namespace log_surgeon {
 MatchedSymbol NonTerminal::m_all_children[cSizeOfAllChildren];
